% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{getSymbol}
\alias{getSymbol}
\alias{getSymbol,Gene-method}
\title{Get the Symbol of a Gene Object}
\usage{
\S4method{getSymbol}{Gene}(object)
}
\arguments{
\item{object}{An object of class Gene.}
}
\value{
The symbol of the gene.
}
\description{
This method returns the symbol of a gene object.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", symbol = "TP53")
getSymbol(gene)
}
