% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{lengthProduct}
\alias{lengthProduct}
\alias{lengthProduct,ProteinCodingGene-method}
\alias{lengthProduct,LongNonCodingRNA-method}
\alias{lengthProduct,MicroRNA-method}
\alias{lengthProduct,Transcript-method}
\title{Get the Length of the Gene Product}
\usage{
lengthProduct(object)
}
\arguments{
\item{object}{An object of a specific gene class.}
}
\value{
The length of the gene product.
}
\description{
This method returns the length of the gene product, depending on the gene type.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", symbol = "TP53", 
                   proteinSequence = "MDM2")
lengthProduct(gene)
}
