% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\docType{class}
\name{MicroRNA-class}
\alias{MicroRNA-class}
\title{MicroRNA Gene Class}
\description{
A class representing a microRNA gene with specific attributes.
}
\section{Slots}{

\describe{
\item{\code{miRNAID}}{A character string representing the miRNA ID.}

\item{\code{seedSequence}}{A character string representing the miRNA seed sequence.}
}}

