% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{setDescription}
\alias{setDescription}
\alias{setDescription,Gene-method}
\title{Set the Description of a Gene Object}
\usage{
setDescription(object, value)
}
\arguments{
\item{object}{An object of class Gene.}

\item{value}{The new description value to assign.}
}
\value{
The modified gene object.
}
\description{
This method sets the description of a gene object.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", description = "Tumor protein p53")
gene <- setDescription(gene, "New Description")
}
