% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{setEnd}
\alias{setEnd}
\alias{setEnd,Gene-method}
\title{Set the End Position of a Gene Object}
\usage{
setEnd(object, value)
}
\arguments{
\item{object}{An object of class Gene.}

\item{value}{The new end position value to assign.}
}
\value{
The modified gene object.
}
\description{
This method sets the end position of a gene object.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", end = 7590856)
gene <- setEnd(gene, 7591000)
}
