% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{setName}
\alias{setName}
\alias{setName,Gene-method}
\title{Set the Name of a Gene Object}
\usage{
setName(object, value)
}
\arguments{
\item{object}{An object of class Gene.}

\item{value}{The new name value to assign.}
}
\value{
The modified gene object.
}
\description{
This method sets the name of a gene object.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", name = "Tumor protein p53")
gene <- setName(gene, "New Gene Name")
}
