% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneClasses.R
\name{setSymbol}
\alias{setSymbol}
\alias{setSymbol,Gene-method}
\title{Set the Symbol of a Gene Object}
\usage{
setSymbol(object, value)
}
\arguments{
\item{object}{An object of class Gene.}

\item{value}{The new symbol value to assign.}
}
\value{
The modified gene object.
}
\description{
This method sets the symbol of a gene object.
}
\examples{
gene <- createGene("protein_coding", ID = "ENSG000001", symbol = "TP53")
gene <- setSymbol(gene, "NEW_SYMBOL")
}
